requirements.txt
attrs==23.2.0
Automat==22.10.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
filelock==3.13.4
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.2.0
jmespath==1.0.1
lxml==5.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.2
parsel==1.9.1
Protego==0.3.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pygsheets==2.0.6
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.11.1
service-identity==24.1.0
six==1.16.0
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.3
